pyproject.toml
[project]
name = "test-fasthtml-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml>=0.5.3",
]
